requirements.txt
streamlit==1.41.*
streamlit-extras==0.4.3
streamlit-float==0.3.5
streamlit-ace==0.1.1
pytest==8.3.1
streamlit-sortables==0.2.0
watchdog==4.0.1
pluggy==1.5.0
python-dotenv==1.0.1
RestrictedPython==7.2
PyGithub~=1.55
pandas==1.5.3
numpy==1.23.2
scipy==1.9.0
PyYAML==6.0.1
tracing==0.1.3
extra_streamlit_components==0.1.71
streamlit_antd_components==0.3.2
gspread~=6.1.2
oauth2client~=4.1.3
openpyxl
bcrypt==4.2.0
locust~=2.31.8
pillow~=9.2.0
qrcode~=7.3.1
matplotlib~=3.9.1
lxml~=5.2.2
openai
networkx
pyvis
graphviz
line-profiler
pyinstaller
pympler
behave
ansi2html
streamlit-javascript
streamlit_js_eval
beartype
